삭제 및 수정 - password 비교 하여 일치 시 작업 수행

+++ src/main/java/com/sparta/week3/controller/BoardController.java
@@ -36,15 +36,23 @@ public class BoardController {
     }
 
     @DeleteMapping("/api/boards/{id}")
-    public Long deleteBoard(@PathVariable Long id) {
-        boardRepository.deleteById(id);
-        return id;
+    public String deleteBoard(@PathVariable Long id, @RequestBody BoardRequestDto requestDto) {
+        Board board = boardRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
+        );
+        if(requestDto.getPassword().equals(board.getPassword())){
+            boardRepository.deleteById(id);
+            return "삭제 성공";
+        }
+        return "비밀번호가 맞지 않습니다";
     }
 
     @PutMapping("/api/boards/{id}")
-    public Long updateBoard(@PathVariable Long id, @RequestBody BoardRequestDto requestDto) {
-        boardService.update(id, requestDto);
-        return id;
+    public String  updateBoard(@PathVariable Long id, @RequestBody BoardRequestDto requestDto) {
+        if(boardService.update(id, requestDto).equals(0L)){
+            return "비밀번호가 맞지 않습니다";
+        }
+        return "수정 성공";
     }
